Console Application
This version uses Visual Studio to create a prototype of the Adnroid app. You can scan any text inside of an image and the words written in it will be than analyzed whether they are present inside our database (we are using a local matrix which mimmics the role of a database). If they are present, the console will print out their definition.